+++ src/app/app.module.ts
@@ -9,7 +9,6 @@ import {ProductsModule} from './products/products.module';
 import {SharedModule} from './shared/shared.module';
 import {AboutComponent} from './layout/components/about/about.component';
 import {CoreModule} from './core/core.module';
-import {OrdersModule} from './orders/orders.module';
 import { PageNotFoundComponent } from './layout/components/page-not-found/page-not-found.component';
 import {HTTP_INTERCEPTORS} from '@angular/common/http';
 import {TimingInterceptor} from './core/interceptors/timing.interceptor';
@@ -33,7 +32,6 @@ const httpInterceptorProviders = [
     ProductsModule,
     CoreModule,
     CartModule,
-    OrdersModule,
     SharedModule,
     AppRoutingModule
   ],